Move tests for problem 1.7 to combined chapter test. Add input processors, add standard assert types. Move repeating code into helper files.
