+++ helpers/assert-types.js
@@ -1,0 +1,4 @@
+module.exports = {
+    equal: 'equal',
+    deepEqual: 'deepEqual',
+};
